requirements.txt
notion-client
langchain
pandas
pdfplumber
Pillow
tqdm
requests
notion-client>=2.0.0
requests>=2.28.0
Pillow>=9.0.0      # optional (for OCR)
pytesseract>=0.3   # optional (for OCR)
PyPDF2
requests
python-dotenv
psutil